+++ benchmarks/utils/src/main/java/fr/unice/yamb/utils/configuration/schema/HeronConfigSchema.java
@@ -7,7 +7,7 @@ public class HeronConfigSchema extends ConfigSchema {
 
     private int maxSpoutPending = 5000;
     private StormDeployment deployment = StormDeployment.local;
-    private int debugFrequency = 0;
+    private float debugFrequency = 0;
 
     public StormDeployment getDeployment() {
         return deployment;
@@ -25,11 +25,11 @@ public class HeronConfigSchema extends ConfigSchema {
         this.maxSpoutPending = maxSpoutPending;
     }
 
-    public int getDebugFrequency() {
+    public float getDebugFrequency() {
         return debugFrequency;
     }
 
-    public void setDebugFrequency(int debugFrequency) {
+    public void setDebugFrequency(float debugFrequency) {
         this.debugFrequency = debugFrequency;
     }
 }
